+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `ScrollableView` has been added
 - `animated` and `completion` parameters can now be passed to methods in `StandardNavigable`
+- `Optional` extension method `let` has been deprecated in favor of `map` and `flatMap`
 
 #### [2.2.1(15)](https://github.com/VakhoKontridze/VCore/releases/download/2.2.1/VCore.xcframework.zip) — *2022 02 14*